Resolved flawed merge, restored _Template page
